+++ src/fmt/mod.rs
@@ -8,6 +8,5 @@ mod indent;
 pub use all::*;
 pub(super) mod all {
     #[doc(inline)]
-    #[cfg(feature = "alloc")]
-    pub use super::indent::indent;
+    pub use super::indent::*;
 }